+++ src/RestApi/RestApi.ts
@@ -5,10 +5,13 @@ import DeviceManagement from "./Device/DeviceManagement";
 import AppletManagement from "./Applet/AppletManagement";
 import OrganizationManagement from "./Organization/OrganizationManagement";
 import FirmwareVersionManagement from "./Firmware/Version/FirmwareVersionManagement";
+import CompanyManagement from './Company/CompanyManagement';
+
 export default class RestApi {
 
 	// Note: We use different authentication here
 	public readonly organization: OrganizationManagement = new OrganizationManagement(this.accountOptions);
+	public readonly company: CompanyManagement = new CompanyManagement(this.accountOptions);
 	public readonly firmwareVersion: FirmwareVersionManagement = new FirmwareVersionManagement(this.accountOptions);
 
 	public readonly timing: TimingManagement = new TimingManagement(this.options);