+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Companies listing and their billing plan settings, organization's subscription type settings
+
 ## [1.2.0] - 2020-01-05
 ### Added
 - Create firmware version with multiple files